added

+++ src/main/java/com/masai/controller/CustomerControl.java
@@ -1,6 +1,9 @@
 package com.masai.controller;
 
 import com.masai.entity.Customer;
+import com.masai.entity.UserType;
+import com.masai.exception.InvalidId;
+import com.masai.exception.NullValueException;
 import com.masai.service.CustomerService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
@@ -14,31 +17,55 @@ public class CustomerControl {
     @Autowired
     private CustomerService customerService;
 
+    @PostMapping("/")
+    public Customer saveCustomerHandler(@RequestBody Customer customer) throws Exception {
+        return customerService.addCustomer(customer,UserType.Customer);
+    }
     @GetMapping("/{id}")
     public Customer viewCustomerHandler(@PathVariable Integer id) {
-        return customerService.viewCustomer(id);
-    }
-
-    @PostMapping("/")
-    public Customer saveCustomerHandler(@RequestBody Customer customer) {
-        return customerService.addCustomer(customer);
+        if(id==null){
+            throw new NullValueException("Invalid URI");
+        }
+        else {
+            return customerService.viewCustomer(id);
+        }
     }
 
     @PutMapping("/{id}/{mobile}/{email}")
     public Customer updateCustomerHandler(@PathVariable Integer id,
                                           @PathVariable String mobile, @PathVariable String email) {
-        Customer customer = customerService.updateCustomer(id, mobile, email);
-        return customer;
+        if(id==null){
+            throw new InvalidId("Invalid URI");
+        }else {
+            Customer customer = customerService.updateCustomer(id, mobile, email);
+            if(customer==null){
+                throw new NullValueException("Customer not found");
+            }
+            return customer;
+        }
+
     }
 
     @DeleteMapping("/{id}")
-    public String removeCustomerHandler(@PathVariable Integer id) {
-        Customer customer = customerService.removeCustomer(id);
-        return "Deleted " + customer;
+    public Customer removeCustomerHandler(@PathVariable Integer id) {
+        if(id==null){
+            throw new InvalidId("Invalid URI");
+        }else {
+            Customer customer = customerService.removeCustomer(id);
+            Customer customer1=customer;
+            if(customer==null){
+                throw new NullValueException("Customer not found");
+            }
+            return  customer1;
+        }
     }
 
     @GetMapping("/all")
     public List<Customer> viewAllCustomersHandler() {
-        return customerService.viewAllCustomers();
+        List<Customer>customers=customerService.viewAllCustomers();
+        if(customers.size()==0){
+            throw new NullValueException("Customer not found");
+        }
+        return customers;
     }
 }
